tests/wpt/tests/selection/selection-direction-on-single-click.html

<!DOCTYPE html>
<html>
<body>
<meta name="author" title="Sean Feng" href="mailto:sean@seanfeng.dev">
<meta name="assert" content="Selection on clicks: direction should return none, forwad, or backward">
<link rel="help" href="https://w3c.github.io/selection-api/#dom-selection-direction">
<link rel="help" href="https://github.com/w3c/selection-api/issues/177">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src='/resources/testdriver-vendor.js'></script>
<div id="container">hello, world</div>
<script>

promise_test(async () => {
    container.innerHTML = 'hello, world';
    const click = new test_driver.Actions()
      .pointerMove(0, 0, container.firstChild)
      .pointerDown()
      .pointerUp()
      .send();
    await click;

    const sel = getSelection();
    assert_equals(sel.anchorNode, container.firstChild);
    assert_equals(sel.anchorOffset, 0);
    assert_equals(sel.focusNode, container.firstChild);
    assert_equals(sel.focusOffset, 0);
    assert_true(sel.isCollapsed);
    assert_equals(sel.direction, 'none');
}, 'direction returns "none" when the selection is collapsed');

</script>
</body>
</html>
