LayoutTests/platform/glib/imported/w3c/web-platform-tests/selection/selection-direction-on-single-click-expected.txt

hello, world

FAIL direction returns "none" when the selection is collapsed assert_equals: expected "none" but got "forward"
